Move `CoreDataFeedStore+FeedImageDataStore` to it's own file
